Refactor color handling and implement Block::Draw function
The color definition has been encapsulated in the `colors.h` file, removing it from `grid.cpp`. The `GetCellColors` function was also removed from the `grid.cpp` and `grid.h` files. Moreover, the `block.cpp` file has been adjusted to include a call to `GetCellColors` in the `Block` constructor and `Draw` function implementation in the `Block` class.